feat: extract FSM spec helpers to helper module
